refactor: rename sectionsubhero instructions to constants (#1598)
