+++ app/components/Home/components/Section/components/SectionSubHero/sectionSubHero.tsx
@@ -1,7 +1,7 @@
 import { SubHeroContent } from "@/components/Home/content";
 import { SectionSubHero as SharedSectionSubHero } from "@repo/shared/views/HomeView/components/Section/components/SectionSubHero/sectionSubHero";
 import { type JSX } from "react";
-import { CTAS, IMAGES, STEPS } from "./instructions";
+import { CTAS, IMAGES, STEPS } from "./constants";
 
 export const SectionSubHero = (): JSX.Element => (
   <SharedSectionSubHero
