+++ app/components/Home/components/Section/components/ga2/SectionSubHero/sectionSubHero.tsx
@@ -1,6 +1,6 @@
 import { SubHeroContent } from "@/components/Home/content/ga2";
 import { type JSX } from "react";
-import { CTAS, IMAGES, STEPS } from "./instructions";
+import { CTAS, IMAGES, STEPS } from "./constants";
 import { StyledSectionSubHero } from "./sectionSubHero.styles";
 
 export const SectionSubHero = (): JSX.Element => (
